Prevent `infdig` error
The SVG output from viz.js is nondeterministic in the order of XML
elements. Ultimately the order of XML elements has no relevance on the
final graphic, so keep a flag and only re-render if the graph has
actually changed.

I was unable to reproduce this in a jasmine test, which is sad.

fix #11