Improve UI - main menu, icons, fewer clicks
Move some global UI out of the editor and into a new menu at the top of
the page.

Hit a really weird error where the uib-dropdown is breaking the
bootstrap3 navbar. It seems to be clearing a float or something, but I
couldn't see where. Managed to screw with floats long enough to make it
reasonable on chrome, but there's something screwy going on there.