First stab at container diagrams
A bunch of messy code for making sub-graphs. Used some lodash templating
to pull the DOT templates out into files, but it's not a clean
transition. Starting putting more into the shell, there might be space
for another directive there. Lots of the stuff added in dot.js should
probably be in model.js.

refs #15
refs #21