New text editor component
Pulls in `angular-ui-codemirror` and gives some config for
s-expressions. This is all a little awkward to use with npm/browserify,
as `angular-ui-codemirror` wants to find `CodeMirror` in the global
namespace. Having to import a bunch of files for side-effects.

Also ports the C4-Lab C4 diagram from JSON to sexp, it's a pretty nice
upgrade in terms of syntax and readability.

refs #29