Add initial redux store, actions, and state
* lots of TODOs in here, ran into trouble blending JS and TS; it's
  easy to use TS from JS, but going the other way gets difficult
* convert some angular component bindings to redux state mapping
* jump through some hoops to make tests happy; TS config with jest
  seems to have some holes

I want to move more UI logic into reducers, but hit some walls with
typescript consuming existing JS code. Started porting the main
`parse` method to typescript, but that's doing a lot of weird stuff
relying on very dynamic typing. Will save that re-write for another
day.

For now, the controllers are doing too much of the work, and
publishing actions to redux as it happens. In some cases I need to get
something from angular into the redux state, and there's more reading
to figure out how to make that happen. Maybe middleware?

Also seeing some weird boilerplate when I want to do some more
formatting in the component based on what state has changed; this is a
question I've had with redux all along: "what should live in the
store, and what should live in the components?"