Move rendering into a "saga"
Experimenting with `redux-saga` here to move work out of the angularjs
controllers. I think it works well, can push small actions to the
redux store as the process commences and trickle updates through the
UI.

The only downside is I had to put some ui-router into my state to
calculate hrefs for the graphviz rendering, that needs to be backed
out to just a function.